Refactored packaging to use mock

+++ yum-plugin-s3-iam.spec
@@ -1,29 +1,36 @@
-Name:     %{name}
-Version:	%{version}
-Release:	%{release}
+Name:		yum-plugin-s3-iam
+Version:	1.0
+Release:	%{__tr_release_num}%{?dist}
 Summary:	Yum package manager plugin for private S3 repositories.
 
-Group:    Application/SystemTools
-License:  Apache License Version 2.0, January 2004
-URL:		  https://github.com/seporaitis/yum-s3-iam
-Source0:	%{name}-%{version}.tar.gz
+Group:		Application/SystemTools
+License:	Apache License Version 2.0, January 2004
+URL:		https://github.com/lattwood/yum-s3-iam
+Source0:	yum-plugin-s3-iam.tar.gz
 
-BuildRoot: %(mktemp -ud %{_tmppath}/%{name}-%{version}-%{release}-XXXXXX)
-BuildArch: noarch
+BuildRoot:	%{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
+BuildArch:	noarch
 
 Requires:	yum
 
 %description
-Yum package manager plugin for private S3 repositories. 
+Yum package manager plugin for private S3 repositories.
 Uses Amazon IAM & EC2 Roles.
 
 %prep
-%setup -q
+rm -rf %{_builddir}/%{name}
+mkdir %{_builddir}/%{name}
+cd %{_builddir}/%{name}
+gzip -dc %{_sourcedir}/yum-plugin-s3-iam.tar.gz | tar -xvvf -
+if [ $? -ne 0 ]; then
+	exit $?
+fi
 
 %build
 
 %install
 rm -rf $RPM_BUILD_ROOT
+cd %{_builddir}/%{name}
 make install DESTDIR=%{buildroot}
 
 %clean
@@ -37,5 +44,8 @@ rm -rf ${RPM_BUILD_ROOT}
 /usr/lib/yum-plugins/s3iam.py
 
 %changelog
+* Mon Oct 20 2014 Logan Attwood <logan@therounds.ca> 1.0-2
+Packaging modifications
+
 * Fri May 31 2013 Matt Jamison <matt@mattjamison.com> 1.0-1
 Initial packaging
